Fix Search Bug (Employee/Shipper)
原先 Employee 及 Shipper 選項必須要選擇,缺少不篩選的選項,此 commit 已修正

+++ eSaleModel/ViewModel/ProductDetail.cs
@@ -24,7 +24,7 @@ namespace eSaleModel.ViewModel
         /// <summary>
         /// 單價
         /// </summary>
-        public List<decimal> UnitPrice { get; set; }
+        public List<int> UnitPrice { get; set; }
         
         /// <summary>
         /// 數量
